added data files

+++ ex2.py
@@ -3,9 +3,13 @@ import numpy as np
 from matplotlib import pyplot as plt
 from skimage.metrics import peak_signal_noise_ratio as psnr, structural_similarity as ssim
 
+import file_io
 from common import calculate_num_frames, pad_frame, split_into_blocks
 
 import logging
+
+from file_io import FileIOHelper
+from input_parameters import InputParameters
 
 logging.basicConfig(format='%(asctime)s.%(msecs)03d %(levelname)-7s [%(filename)s:%(lineno)d] %(message)s',
     datefmt='%H:%M:%S',
@@ -30,13 +34,21 @@ def read_y_component(file_path, width, height, num_frames):
 
 
 # Function to save Y-only frames to individual files
-def save_y_frames(input_file, output_file, width, height):
-    num_frames = calculate_num_frames(input_file, width, height)
-    if os.path.exists(output_file):
+def save_y_frames_to_file(params : InputParameters, frames_to_extract=None):
+    if params.yuv_file is None:
+        guessed_yuv_file_name = FileIOHelper(params).get_yuv_file_name()
+        logger.info(f"No yuv file provided, assuming yuv = [{guessed_yuv_file_name}] ")
+        params.yuv_file = guessed_yuv_file_name
+
+    input_file = params.yuv_file
+
+    output_file= params.y_only_file
+    num_frames = frames_to_extract if frames_to_extract else calculate_num_frames(params.yuv_file, params.width, params.height)
+    if os.path.exists(params.yuv_file):
         logger.info(f"y only {output_file} already exists. skipping...")
         return
     with open(input_file, 'rb') as f_in, open(output_file, 'wb') as f_out:
-        for frame_index, y_plane in enumerate(read_y_component(input_file, width, height, num_frames)):
+        for frame_index, y_plane in enumerate(read_y_component(input_file, params.width, params.height, num_frames)):
             f_out.write(y_plane.tobytes())
 
 
@@ -56,16 +68,24 @@ def reconstruct_frame_from_blocks(blocks, frame_shape, block_size):
 
 
 # Function to process Y-only files and split them into blocks
-def process_y_frames(input_file, width, height, block_sizes):
+def process_y_frames(params : InputParameters, block_sizes):
+    input_file = params.y_only_file
+    width = params.width
+    height = params.height
     logger.info(f"Processing file: {input_file}")
-    file_prefix = os.path.splitext(input_file)[0]
+    file_io_h = FileIOHelper(params)
 
     file_handles = {}
 
     # Open output files dynamically based on block sizes
     for block_size in block_sizes:
-        file_name = f'{file_prefix}-{block_size}block.y'
+        file_name = file_io_h.get_file_name_wo_identifier(f'{block_size}b.y')
+        if os.path.exists(file_name):
+            continue
         file_handles[block_size] = open(file_name, 'wb')
+
+    if len(file_handles) < 1:
+        return
 
     y_size = width * height
 
@@ -154,19 +174,22 @@ def plot_quality_metrics(block_sizes, psnr_values, ssim_values):
     plt.legend()
 
     plt.tight_layout()
-    plt.show()
+    # plt.show()
 
 
-def main(input_file, width, height):
-    file_prefix = os.path.splitext(input_file)[0]
-    y_only_file = f'{file_prefix}.y'
-    save_y_frames(input_file, y_only_file, width, height)
+def main(params: InputParameters):
+    width = params.width
+    height = params.height
+    file_io = FileIOHelper(params)
+
+    y_only_file = params.y_only_file
+    save_y_frames_to_file(params)
 
     block_sizes = [1, 2, 8, 16, 64]  # Block sizes to process
-    process_y_frames(y_only_file, width, height, block_sizes)
+    process_y_frames(params, block_sizes)
 
     for block_size in block_sizes:
-        averaged_file = f'{file_prefix}-{block_size}block.y'
+        averaged_file = file_io.get_file_name_wo_identifier(f'{block_size}b.y')
         average_psnr, average_ssim = calculate_psnr_ssim(y_only_file, averaged_file, width, height)
         # Store results for plotting
         if 'psnr_results' not in locals():
